Refactor Named Pipe handling to improve code organization and error handling
